+++ kaithem/src/docs/changes.md
@@ -10,6 +10,8 @@ Change Log
 * :technologist: Clear cache in browser on version change. No more version UUIDs in URLs
 * :technologist: Lots of linter warnings that apparently just showed up in an eslint update fixed.
 * :technologist: Cleaner layout for resource metadata. Backwards compatible, old stuff is auto-upgraded.
+* :technologist: mutable_copy_resource in kaithem.api.modules
+
 
 
 ### 0.92.1